Replace tree scaffolding using functions with tree scaffolding using relations
The old scaffolding is exponential, so it is only feasible for small trees
(and preferrable in that case, since using functions allows some
rules to be embedded in linear time and space, while using relations cubic
embeddings are necessary. This is costly but not unfeasible for a
couple of hundreds of nodes, which is beyond the reach of the old embedding).